added navbar component to be re-used on all pages
added the navbar to the profile page
